+++ lib/slimcop/slim_offense_collector.rb
@@ -25,7 +25,7 @@ module Slimcop
         ).call.map do |rubocop_offense|
           Offense.new(
             file_path: @file_path,
-            offset: snippet[:begin_],
+            offset: snippet[:offset],
             rubocop_offense: rubocop_offense,
             source: @source
           )